Изменения на MsgToSessionService.
