+++ src/main/java/ru/aviaj/messagesystem/message/sessionservice/MsgLogin.java
@@ -1,22 +1,24 @@
-package ru.aviaj.messagesystem.message.authcontroller;
+package ru.aviaj.messagesystem.message.sessionservice;
 
 import ru.aviaj.database.exception.DbException;
 import ru.aviaj.messagesystem.Address;
 import ru.aviaj.service.SessionService;
 
-public class MsgLogout extends MsgToAuthController {
+public class MsgLogin extends MsgToSessionService {
 
     private String sessionId;
+    private long userId;
 
-    public MsgLogout(Address from, Address to, String sessionId, long userId) {
+    public MsgLogin(Address from, Address to, String sessionId, long userId) {
         super(from, to);
         this.sessionId = sessionId;
+        this.userId = userId;
     }
 
     @Override
     public void exec(SessionService sessionService) {
         try {
-            sessionService.removeSession(this.sessionId);
+            sessionService.addSession(this.sessionId, this.userId);
         } catch (DbException e) {
             //Error back
         }
